Use parallel sequence ops to analyze feed messages
This removes the 'analyze()' function which was used per-feed to find
the longest message in a feed, the most retweets in a feed, etc., with
an explicit outer loop over the feeds.

The operations were reimplemented as fold/count/max operations on
parallel sequences.

This in turn avoids using the 'Accumulator' class, which used an
'AtomicReference' to keep track of intermediate results and which was
only used in a multi-threaded context (and thus warrant the use of the
atomic) by 'mostRechirpsInAllFeeds'. All the other functions called
'analyze()' in thread-local context.